requirements.txt
django==1.6.7
pillow
django-webodt
pytz
gunicorn
